requirements.txt
aiomysql==0.2.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
fastapi==0.115.12
h11==0.16.0
idna==3.10
iso8601==2.1.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.1
pypika-tortoise==0.6.1
python-multipart==0.0.20
pytz==2025.2
sniffio==1.3.1
starlette==0.46.2
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
